requirements.txt
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cssselect2==0.7.0
defusedxml==0.7.1
dotmap==1.3.30
enum-compat==0.0.3
fontawesome==5.10.1.post1
geographiclib==2.0
geopy==2.3.0
idna==3.4
meshtastic==2.1.8
numpy==1.25.1
pexpect==4.8.0
Pillow==8.4.0
pillow-svg==0.1.0
protobuf==4.23.3
ptyprocess==0.7.0
pycparser==2.21
pygatt==4.0.5
pypandoc==1.11
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0
requests==2.31.0
RPi.GPIO==0.7.1
spidev==3.6
tabulate==0.9.0
timeago==1.0.16
tinycss2==1.2.1
urllib3==2.0.3
webencodings==0.5.1
